use byte[] in CTAEvent as in this moment we read directly the byte data

+++ src/main/java/streams/cta/CTAEvent.java
@@ -15,9 +15,9 @@ public class CTAEvent implements Serializable {
 	private static final long serialVersionUID = -595319426051864811L;
 
 	protected final int numberOfPixels;
-	protected final short[] data;
+	protected final byte[] data;
 
-	public CTAEvent(int numPixels, short[] data) {
+	public CTAEvent(int numPixels, byte[] data) {
 		this.numberOfPixels = numPixels;
 		this.data = data;
 	}
@@ -26,7 +26,7 @@ public class CTAEvent implements Serializable {
 		return numberOfPixels;
 	}
 
-	public short[] values() {
+	public byte[] values() {
 		return data;
 	}
 }
